move common-use functions to core module
