+++ lib/spack/spack/cmd/find.py
@@ -204,16 +204,6 @@ def display_env(env, args, decorator):
         print()
 
 
-def filter_loaded_specs(specs):
-    """Filter a list of specs returning only those that are currently loaded."""
-    hashes = os.environ.get(uenv.spack_loaded_hashes_var, '').split(':')
-    return [x for x in specs if x.dag_hash() in hashes]
-
-
-def print_how_many_pkgs(specs, pkg_type=""):
-    tty.msg("%s" % plural(len(specs), pkg_type + " package"))
-
-
 def find(parser, args):
     if args.bootstrap:
         bootstrap_store_path = spack.bootstrap.store_path()
@@ -251,7 +241,7 @@ def _find(parser, args):
         results = [x for x in results if x.name in packages_with_tags]
 
     if args.loaded:
-        results = filter_loaded_specs(results)
+        results = spack.cmd.filter_loaded_specs(results)
 
     # Display the result
     if args.json:
@@ -263,7 +253,7 @@ def _find(parser, args):
 
         if sys.stdout.isatty() and args.groups:
             pkg_type = "loaded" if args.loaded else "installed"
-            print_how_many_pkgs(results, pkg_type)
+            spack.cmd.print_how_many_pkgs(results, pkg_type)
 
         cmd.display_specs(
             results, args, decorator=decorator, all_headers=True)
